Adds Error override for SisbaseResult
This override will be used on all cases that old Logger.Error were
used to report an inner error from a SisbaseResult.

The intent of adding this function is to make it impossible
for errors to go unreported, since all this will basically be called
any time an SisbaseResult instance can error.

It also has a obligatory source param, so we can indicate the place that the
error originated, making it easy to pinpoint and fix bugs.